@@ -5,10 +5,10 @@
     "@rails/webpacker": "^5.4.3",
     "@sfgrp/sled": "^1.0.3",
     "@sfgrp/svg-radial-menu": "^1.0.7",
-    "axios": "^0.25.0",
-    "chart.js": "^3.7.0",
+    "axios": "^0.26.0",
+    "chart.js": "^3.7.1",
     "d3": "^7.3.0",
-    "dompurify": "^2.3.5",
+    "dompurify": "^2.3.6",
     "easymde": "^2.16.1",
     "exif-js": "^2.3.0",
     "json2csv": "^5.0.6",
@@ -18,7 +18,7 @@
     "pdfjs-dist": "2.3.200",
     "qs": "^6.10.3",
     "v-hotkey": "^0.9.0",
-    "vue": "^3.2.29",
+    "vue": "^3.2.31",
     "vue-handy-scroll": "^3.0.1",
     "vue-hotkeys": "0.0.1",
     "vue-tippy": "^6.0.0-alpha.32",
@@ -28,18 +28,18 @@
     "webpack-cli": "^3.3.12"
   },
   "devDependencies": {
-    "@vue/compiler-sfc": "^3.2.29",
+    "@vue/compiler-sfc": "^3.2.31",
     "babel-plugin-syntax-dynamic-import": "^6.18.0",
     "babel-preset-stage-0": "^6.24.1",
     "css-loader": "^5.2.7",
     "css-minimizer-webpack-plugin": "^2.0.0",
-    "eslint": "^8.7.0",
+    "eslint": "^8.9.0",
     "eslint-config-standard": "^16.0.3",
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^6.0.0",
     "eslint-plugin-standard": "^4.1.0",
-    "eslint-plugin-vue": "^8.3.0",
+    "eslint-plugin-vue": "^8.4.1",
     "increase-memory-limit": "^1.0.7",
     "mini-css-extract-plugin": "^1.6.2",
     "raw-loader": "^4.0.2",
@@ -48,7 +48,7 @@
     "sass-loader": "^10.2.0",
     "stylus": "^0.54.8",
     "stylus-loader": "^4.3.3",
-    "vue-loader": "^16.5.0",
+    "vue-loader": "^17.0.0",
     "webpack-dev-server": "^3.11.2"
   },
   "scripts": {
