@@ -75,7 +75,10 @@
         </tr>
       </thead>
       <tbody>
-        <tr v-for="(item, index) in list">
+        <tr
+          v-for="(item, index) in list"
+          :key="item.id"
+        >
           <td><span v-html="item.document.object_tag" /></td>
           <td>
             <input
@@ -100,9 +103,17 @@
                   x-small
                   name="download"/>
               </v-btn>
-              <span
-                class="button circle-button btn-delete"
-                @click="removeItem(item)"/>
+              <v-btn
+                circle
+                class="circle-button"
+                color="destroy"
+                @click="confirmDelete(item)"
+              >
+                <v-icon
+                  name="trash"
+                  x-small
+                />
+              </v-btn>
             </div>
           </td>
         </tr>
@@ -122,7 +133,11 @@ import VIcon from 'components/ui/VIcon/index'
 import VBtn from 'components/ui/VBtn/index'
 
 export default {
-  mixins: [CRUD, annotatorExtend],
+  mixins: [
+    CRUD,
+    annotatorExtend
+  ],
+
   components: {
     RadialAnnotator,
     Autocomplete,
@@ -131,12 +146,14 @@ export default {
     VBtn,
     VIcon
   },
+
   computed: {
     validateFields () {
       return this.documentation.document_id
     }
   },
-  data: function () {
+
+  data () {
     return {
       display: 0,
       optionList: ['drop', 'pick', 'pinboard'],
@@ -156,9 +173,11 @@ export default {
       }
     }
   },
+
   created () {
     this.$options.components['RadialAnnotator'] = RadialAnnotator
   },
+
   methods: {
     newDocumentation () {
       return {
@@ -166,20 +185,24 @@ export default {
         annotated_global_entity: decodeURIComponent(this.globalId)
       }
     },
+
     createNew () {
       this.create('/documentation', { documentation: this.documentation }).then(response => {
         this.list.push(response.body)
         this.documentation = this.newDocumentation()
       })
     },
-    success: function (file, response) {
+
+    success (file, response) {
       this.list.push(response)
       this.$refs.figure.removeFile(file)
     },
-    sending: function (file, xhr, formData) {
+
+    sending (file, xhr, formData) {
       formData.append('documentation[annotated_global_entity]', decodeURIComponent(this.globalId))
       if (this.isPublic) { formData.append('documentation[document_attributes][is_public]', this.isPublic) }
     },
+
     changeIsPublicState (index, documentation) {
       const data = {
         id: documentation.document_id,
@@ -188,7 +211,13 @@ export default {
       this.update(`/documents/${data.id}.json`, { document: data }).then(response => {
         this.list[index].is_public = response.body.is_public
       })
-    }
+    },
+
+    confirmDelete (item) {
+      if (window.confirm("You're trying to delete this record. Are you sure want to proceed?")) {
+        this.removeItem(item)
+      }
+    },
   }
 }
 </script>
