@@ -1,8 +1,12 @@
 import { createApp } from 'vue'
-import App from './tempApp.vue'
+import { newStore } from './store/store'
+import MatrixRowCoderRequest from './request/MatrixRowCoderRequest'
+import App from './App.vue'
 
 function init () {
   const app = createApp(App)
+  const store = newStore(new MatrixRowCoderRequest())
+  app.use(store)
   app.mount('#matrix_row_coder')
 }
 
