@@ -3,12 +3,8 @@ import DescriptorTypes from '../helpers/DescriptorTypes'
 import ComponentNames from '../helpers/ComponentNames'
 import makeEmptyObservationsFor from '../helpers/makeEmptyObservationsFor'
 
-export default function ({ commit, state }, args) {
-  const {
-    rowId,
-    otuId
-  } = args
-  return state.request.getMatrixRow(rowId, otuId)
+export default function ({ commit, state }, rowId) {
+  return state.request.getMatrixRow(rowId)
     .then(response => {
       const descriptors = response.descriptors.map(transformDescriptorForViewmodel)
       commit(MutationNames.SetDescriptors, descriptors)
@@ -45,7 +41,6 @@ function makeBaseDescriptor (descriptorData) {
     globalId: descriptorData.global_id,
     description: getDescription(descriptorData),
     type: getComponentNameForDescriptorType(descriptorData),
-    isZoomed: false,
     isUnsaved: false,
     needsCountdown: false,
     isSaving: false,
