@@ -23,7 +23,7 @@ class DatasetRecord::DarwinCore::Taxon < DatasetRecord::DarwinCore
   def import(dwc_data_attributes = {})
     super
     begin
-      DatasetRecord.transaction do
+      DatasetRecord.transaction(requires_new: true) do
         self.metadata.delete('error_data')
 
         nomenclature_code = get_field_value('nomenclaturalCode')&.downcase&.to_sym || import_dataset.default_nomenclatural_code
@@ -157,6 +157,7 @@ class DatasetRecord::DarwinCore::Taxon < DatasetRecord::DarwinCore
                 end
 
                 if (rank_in_type = ORIGINAL_COMBINATION_RANKS[rank&.downcase&.to_sym])
+                  taxon_name.save!
                   TaxonNameRelationship.find_or_create_by!(type: rank_in_type, subject_taxon_name: ancestor_protonym, object_taxon_name: taxon_name)
                 end
               end
